refactor: add explicit underlying types to enums
* Set the underlying type of `ffa_error` to `uint32_t`. This reduces the
  need for explicit casts in most cases.
* Set the underlying types for other enums to `uint32_t`, where they are
  specified as such in the FF-A specification (`ffa_version`,
  `ffa_feature_id`, `ffa_mem_perm`, `ffa_framework_msg_func`).
* Set the underlying types of `ffa_data_access`,
  `ffa_instruction_access`, `ffa_memory_type`,
  `ffa_memory_cacheability`, `ffa_memory_shareability`,
  `ffa_memory_security` to `uint8_t`. This allows them to be stored
  directly in the `ffa_data_access_t` and `ffa_memory_attributes_t`
  structs without increasing their size.

+++ src/api.c
@@ -2408,7 +2408,7 @@ struct ffa_value api_ffa_version(struct vcpu *current,
 	static_assert(sizeof(enum ffa_version) == 4,
 		      "enum ffa_version must be 4 bytes wide");
 
-	const struct ffa_value error = {.func = (uint32_t)FFA_NOT_SUPPORTED};
+	const struct ffa_value error = {.func = FFA_NOT_SUPPORTED};
 	struct vm_locked current_vm_locked;
 
 	uint16_t compiled_major = ffa_version_get_major(FFA_VERSION_COMPILED);
@@ -2456,7 +2456,7 @@ struct ffa_value api_ffa_version(struct vcpu *current,
 	current_vm_locked.vm->ffa_version = requested_version;
 	vm_unlock(&current_vm_locked);
 
-	return (struct ffa_value){.func = (uint32_t)FFA_VERSION_COMPILED};
+	return (struct ffa_value){.func = FFA_VERSION_COMPILED};
 }
 
 /**
